+++ src/blockchain/mod.rs
@@ -1,7 +1,0 @@
-mod address;
-mod contract;
-mod ipfs;
-
-pub use address::{Eth, Sol};
-pub use contract::CustomContract;
-pub use ipfs::normalize_ipfs_url;