+++ src/graphql/token/metadata/common/mod.rs
@@ -1,5 +1,7 @@
 mod attrs;
 mod custom_mime;
+mod media;
 
 pub use attrs::Attributes;
 pub use custom_mime::Mime;
+pub use media::Media;