+++ CHANGELOG.md
@@ -6,3 +6,32 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.0] - 21/10/2022
+
+### Added
+
+- Github pages Docs.rs documentation
+- Fast and concurrent token fetching and processing
+- Basic Query fields caching
+- Common metadata data fields
+- Possibilities of extending Web3 Contract
+- Possibilities of extending token metadata
+- `OpenSea` example of extending token metadata
+- Pagination(with offsets and limits)
+- IPFS support
+- Full Documentation
+- New token getters:
+  - All tokens
+  - One token
+  - Page of tokens
+
+## [0.1.0] - 24/09/2022
+
+### Added
+
+- Web3 Contract implementation
+- GraphQL `token` field
+- Basic Token schema
+- Basic Ethereum address validation
+- Functions for starting `actix_web` server, `GraphQL` endpoint and `GraphiQL` playground